newUser.htmlにたいしてnewUserStyle.cssがリンクしなくなった

+++ src/main/java/com/example/demo/repository/PostRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.Post;
 
-public interface PostRepository extends JpaRepository<Post, Long> {
-	
+public interface PostRepository extends JpaRepository<Post, Integer> {
+//	List<Post>findByUserIdAndMessage(Integer id,Integer userid, String message );
+
 	}
